requirements.txt
chromadb==0.4.22
gradio>=4.44.1
sentence-transformers>=4.1.0
pandas>=2.2.3
numpy>=1.26.0
python-dotenv==1.0.0
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
pypdf2==3.0.1
python-docx==1.1.0
audioop-lts
